+++ src/YaraParser/TransitionBasedSystem/Parser/BeamScorerThread.java
@@ -40,7 +40,7 @@ public class BeamScorerThread implements Callable<ArrayList<BeamElement>> {
         ArrayList<BeamElement> elements = new ArrayList<BeamElement>(dependencyRelations.size() * 2 + 3);
 
         State currentState = configuration.state;
-        float prevScore = configuration.score;
+        double prevScore = configuration.score;
 
         boolean canShift = ArcEager.canDo(Actions.Shift, currentState);
         boolean canReduce = ArcEager.canDo(Actions.Reduce, currentState);
@@ -49,30 +49,30 @@ public class BeamScorerThread implements Callable<ArrayList<BeamElement>> {
         Object[] features = FeatureExtractor.extractAllParseFeatures(configuration, featureLength);
 
         if (canShift) {
-            float score = classifier.shiftScore(features, isDecode);
-            float addedScore = score + prevScore;
+            double score = classifier.shiftScore(features, isDecode);
+            double addedScore = score + prevScore;
             elements.add(new BeamElement(addedScore, b, 0, -1));
         }
         if (canReduce) {
-            float score = classifier.reduceScore(features, isDecode);
-            float addedScore = score + prevScore;
+            double score = classifier.reduceScore(features, isDecode);
+            double addedScore = score + prevScore;
             elements.add(new BeamElement(addedScore, b, 1, -1));
 
         }
 
         if (canRightArc) {
-            float[] rightArcScores = classifier.rightArcScores(features, isDecode);
+            double[] rightArcScores = classifier.rightArcScores(features, isDecode);
             for (int dependency : dependencyRelations) {
-                float score = rightArcScores[dependency];
-                float addedScore = score + prevScore;
+                double score = rightArcScores[dependency];
+                double addedScore = score + prevScore;
                 elements.add(new BeamElement(addedScore, b, 2, dependency));
             }
         }
         if (canLeftArc) {
-            float[] leftArcScores = classifier.leftArcScores(features, isDecode);
+            double[] leftArcScores = classifier.leftArcScores(features, isDecode);
             for (int dependency : dependencyRelations) {
-                float score = leftArcScores[dependency];
-                float addedScore = score + prevScore;
+                double score = leftArcScores[dependency];
+                double addedScore = score + prevScore;
                 elements.add(new BeamElement(addedScore, b, 3, dependency));
             }
         }
